remove atlas project and move infra into CLI
