+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
-  "atlas",
   "cli",
   "graphql-parser"
 ]
