--- atlas/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "atlas"
-version = "0.1.0"
-authors = ["jbaxleyiii <james@apollographql.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-atty = "0.2"
-console = "0.6.1"
-log = { version = "0.4", features = ["std"] }
-textwrap = "0.11.0"
-term_size = "0.3.0"
-serde = "1.0"
-failure = "*"
-dirs = "2.0"
