--- atlas/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod log;
-pub mod style;
-pub mod errors;
-pub mod layout;
