+++ cli/Cargo.toml
@@ -7,13 +7,16 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-atlas = { path = "../atlas" }
-structopt = "*"
-serde = "1.0"
+atty = "0.2"
+console = "0.6.1"
+dirs = "2.0"
 failure = "*"
 graphql-parser = { path = "../graphql-parser" }
 log = { version = "0.4", features = ["std"] }
-dirs = "2.0"
+serde = "1.0"
+structopt = "*"
+term_size = "0.3.0"
+textwrap = "0.11.0"
 
 [dev-dependencies]
 assert_cmd = "0.10"
