+++ cli/src/log.rs
