+++ cli/tests/cli.rs
